Refactor badge parsing/serialization/typing for reuse.
There are subtle differences between object stores, user object store instance models, and object store templates in the downstream "bring-your-own-object-store" branch so breaking this stuff out like this should allow for cleaner diffs and reuse.
